Adding customization property for the HttpClient on ConnectionRequestTimeout
The ConnectionRequestTimeout property is needed to initalize correctly the timeout of a Request of the HTTPClient, but the default value setted is totaly sufficient so there isn't need to set it in portlet.xml preferences.
This commit introduce the possibility to configure it from preference properties in case of need.

With more development you can add other properties to be able to set :
	- The RequestConfig : https://hc.apache.org/httpcomponents-client-ga/httpclient/apidocs/org/apache/http/client/config/RequestConfig.Builder.html
	- The HttpClient with HttpClientBuilder : https://hc.apache.org/httpcomponents-client-ga/httpclient/apidocs/org/apache/http/impl/client/HttpClientBuilder.html

Also with this new HttpClient some code could be rebuilded (interceptors, keepalived, ssl, proxy, etc.)